(TRAINTECH-1530) Add Support For API V3
TRAINTECH-1530 #resolved #time 2d
This is not full support for the new graphql API from LearnDot.
The new API does not yet support the necessary objects that we need
to build on top of.  This is just the proof of concept that we can
use the new API once it is updated by LearnDot.  Includes a fixed
schema file and very simple query examples.